Separate capture rate from led refresh rate
For smoothing to work better. Skip frame capture if it has not been long
enough, so as to be able to update leds due to smoothing

+++ src/config.rs
@@ -112,7 +112,7 @@ pub struct Transform {
 struct Smoothing {
 	pub type_: String,
 	pub time_ms: u32,
-	pub update_frequency: f32
+	pub update_frequency: f64
 }
 impl Decodable for Smoothing {
 	fn decode<D: Decoder>(decoder: &mut D) -> Result<Smoothing, D::Error> {
